Move the linkding import under an lz-cli subcommand
